♻️ seperate different types of stratigies
