+++ pyrb/controller.py
@@ -4,7 +4,7 @@ from pyrb.brokerage.base.client import TradeMode
 from pyrb.brokerage.base.order_manager import Order, OrderStatus
 from pyrb.brokerage.context import RebalanceContext, create_rebalance_context
 from pyrb.service.rebalance import Rebalancer
-from pyrb.service.strategy import HoldingPortfolioRebalancingStrategy
+from pyrb.service.strategy.direct_rebalance import HoldingPortfolioRebalancingStrategy
 
 app = typer.Typer()
